First signs of life using EF Core

+++ Api/Database/OrderProductEntity.cs
@@ -2,8 +2,8 @@ namespace Api.Database;
 
 public class OrderProductEntity
 {
-    public long OrderId { get; set; }
-    public long ProductId { get; set; }
+    public int OrderId { get; set; }
+    public int ProductId { get; set; }
     public double Price { get; set; }
     public int Quantity { get; set; }
 }
